subdomains fixed

+++ database/migrations/2025_11_10_192202_create_social_security_items_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('social_security_items')) {
+            return;
+        }
+
         Schema::create('social_security_items', function (Blueprint $table) {
             $table->id();
             $table->foreignId('social_security_id')->constrained('social_security')->onDelete('cascade');
